chore: renamed f() to gaussJ()

+++ src/GaussJordan.java
@@ -3,9 +3,9 @@ import java.util.List;
 import java.util.Scanner;
 
 public class GaussJordan {
-    public static float[] gaussJ(Matrix m, Scanner in){
-        float[] err = {-405};
-        float[] solusi = new float[m.col-1];
+    public static double[] gaussJ(Matrix m, Scanner in){
+        double[] err = {-405};
+        double[] solusi = new double[m.col-1];
         boolean[] isFree = new boolean[m.col-1];
         boolean unsolv=false;
         boolean hasFreeVar = false;
@@ -24,7 +24,7 @@ public class GaussJordan {
             }
 
             // Step 2: Normalize the row so that pivot becomes 1
-            float pivot = m.elmt(i, i);
+            double pivot = m.elmt(i, i);
             if (pivot != 0) {
                 m.multRowByK(i, 1 / pivot);
             }
@@ -32,7 +32,7 @@ public class GaussJordan {
             // Step 3: Eliminate all elements below the pivot in the current column
             for (int j = i + 1; j < N; j++) {
                 if(!m.isZeroRow(j)){
-                    float factor = -m.elmt(j, i);
+                    double factor = -m.elmt(j, i);
                     m.addRow(j, i, factor);
                 }
                 
@@ -41,11 +41,11 @@ public class GaussJordan {
 
         /* Eliminasi Gauss-Jordan */
         for(int i=N-1;i>=0;i--){
-            float pivot = m.elmt(i,i);
+            double pivot = m.elmt(i,i);
 
             for(int j=i-1;j>=0;j--){
                 if(!m.isZeroRow(j)){
-                    float factor = -m.elmt(j, i);
+                    double factor = -m.elmt(j, i);
                     m.addRow(j, i, factor);
                 }
             }
@@ -59,7 +59,7 @@ public class GaussJordan {
             else if(m.isZeroRow(i)){
                 freeVar.add(i); //variabel bebas
             }
-            float sum=0;
+            double sum=0;
             for(int j=i+1;j<m.col-1;j++){
                 sum+=m.elmt(i,j)*solusi[j];
             }
@@ -134,7 +134,7 @@ public class GaussJordan {
     
 
     public static void main(String[] args) {
-        float[] ans;
+        double[] ans;
         Scanner in = new Scanner(System.in);
         
         Matrix mat = null;
